config: drop unneeded ._load subpackage
