+++ src/lenskit/config/__init__.py
@@ -13,6 +13,7 @@ from __future__ import annotations
 import warnings
 from contextlib import AbstractContextManager, contextmanager
 from contextvars import ContextVar
+from os import PathLike
 from pathlib import Path
 
 from pydantic_settings import TomlConfigSettingsSource
@@ -30,8 +31,6 @@ from lenskit.schemas.settings import (
     RandomSettings,
     TuneSettings,
 )
-
-from ._load import locate_configuration_root
 
 __all__ = [
     "lenskit_config",
@@ -197,3 +196,43 @@ def _load_settings[C](cfg_dir: Path | None, settings_cls: type[C]) -> C:
     obj = LenskitFileSettings()
     obj.finish_setup()
     return obj  # type: ignore
+
+
+def locate_configuration_root(
+    *,
+    cwd: Path | str | PathLike[str] | None = None,
+    abort_at_pyproject: bool = True,
+    abort_at_gitroot: bool = True,
+) -> Path | None:
+    """
+    Search for a configuration root containing a ``lenskit.toml`` file.
+
+    This searches for a ``lenskit.toml`` file, beginning in the current working
+    directory (or the alternate ``cwd`` if provided), and searching upward until
+    one is found.  Search stops if a ``pyproject.toml`` file or ``.git``
+    directory is found without encountering ``lenskit.toml``.
+    """
+
+    if cwd is None:
+        cwd = Path()
+    elif not isinstance(cwd, Path):
+        cwd = Path(cwd)
+    cwd = cwd.resolve()
+
+    log = _log.bind(cwd=str(cwd))
+    log.debug("searching for lenskit.toml")
+    while cwd is not None:
+        log.debug("checking if lenskit.toml exists", dir=str(cwd))
+        if (cwd / "lenskit.toml").exists():
+            return cwd
+
+        if abort_at_pyproject and (cwd / "pyproject.toml").exists():
+            break
+
+        if abort_at_gitroot and (cwd / ".git").exists():
+            break
+
+        if cwd.parent == cwd:
+            break
+        else:
+            cwd = cwd.parent
